+++ core-network/src/main/java/com/hmoa/core_network/service/CommunityComment/CommunityCommentServiceImpl.kt
@@ -4,7 +4,7 @@ import com.hmoa.core_model.request.CommunityCommentDefaultRequestDto
 import com.hmoa.core_model.response.CommunityCommentAllResponseDto
 import com.hmoa.core_model.response.CommunityCommentDefaultResponseDto
 import com.hmoa.core_model.response.DataResponseDto
-import com.hmoa.core_network.HttpClientProvider
+import com.hmoa.core_network.service.HttpClientProvider
 import io.ktor.client.call.*
 import io.ktor.client.request.*
 import io.ktor.util.*
