+++ core-network/src/main/java/com/hmoa/core_network/service/Fcm/FcmServiceImpl.kt
@@ -2,7 +2,7 @@ package corenetwork.Fcm
 
 import com.hmoa.core_model.request.FCMTokenSaveRequestDto
 import com.hmoa.core_model.response.DataResponseDto
-import com.hmoa.core_network.HttpClientProvider
+import com.hmoa.core_network.service.HttpClientProvider
 import io.ktor.client.call.*
 import io.ktor.client.request.*
 import io.ktor.util.*
