+++ core-network/src/main/java/com/hmoa/core_network/service/Note/NoteServiceImpl.kt
@@ -1,7 +1,7 @@
 package corenetwork.Note
 
 import com.hmoa.core_model.response.DataResponseDto
-import com.hmoa.core_network.HttpClientProvider
+import com.hmoa.core_network.service.HttpClientProvider
 import io.ktor.client.call.*
 import io.ktor.client.request.*
 import io.ktor.http.*
